map perms; more option toggles identified
Converted map permissions to be stored as signed in-memory, as error message strings in MegaloEdit describe these IDs as lying in the range "(-32k, +32k)".

Identified the following Option Toggles by reverse-engineering MegaloEdit: Perfection Enabled; Team Lives per Round; Proximity Voice; Don't Team-Restrict Voice Chat; Allow Dead Players to Talk; Shortcuts on Map; Red Powerup Duration; Blue Powerup Duration; Custom Powerup Duration.

+++ native/src/ReachVariantTool/game_variants/components/map_permissions.h
@@ -17,20 +17,20 @@ class ReachMapPermissions {
       static constexpr int max_count = 32;
       //
    public:
-      std::vector<uint16_t> mapIDs;
+      std::vector<int16_t> mapIDs;
       cobb::bitnumber<1, reach::map_permission_type> type;
       //
       void read(cobb::ibitreader& stream) noexcept {
          size_t count = stream.read_bits<size_t>(6);
          this->mapIDs.resize(count);
          for (size_t i = 0; i < count; i++)
-            this->mapIDs[i] = stream.read_bits<uint16_t>(16);
+            this->mapIDs[i] = (int16_t) stream.read_bits<uint16_t>(16);
          this->type.read(stream);
       }
       void write(cobb::bitwriter& stream) const noexcept {
          stream.write(this->mapIDs.size(), 6);
-         for (uint16_t id : this->mapIDs)
-            stream.write(id);
+         for (int16_t id : this->mapIDs)
+            stream.write((uint16_t)id);
          this->type.write(stream);
       }
       //
